refactored main structure
