+++ .travis.yml
@@ -1,4 +1,5 @@
 language: java
-script: mvn clean package
 jdk:
   - oraclejdk8
+script: mvn clean test
+#after_script: mvn clean test -Pintegration
