+++ src/main/java/com/csoft/muon/dao/DataSourceFactory.java
@@ -1,4 +1,4 @@
-package com.csoft.muon.lib;
+package com.csoft.muon.dao;
 
 import java.sql.SQLException;
